Init templates filling in code generator

+++ Sources/ReferenceExtractor/ReferenceExtractor.swift
@@ -21,6 +21,9 @@ import Common
 ///
 /// `ReferenceExtractor` is proof to reference cycles
 ///
+/// **ATTENTION**
+/// Doesn't exclude local references. SO if ypu have reference which is local (wihtout file path before `#`)
+/// Extractor doesn't return it in result array.
 ///
 /// **WARNING**
 /// Isn't thread-safe!
@@ -71,7 +74,9 @@ extension ReferenceExtractor {
     /// Doesn't return link on `rootSpecPath`
     /// If you need it you shoul do it by yourself
     public func extract() throws -> (uniqRefs: [String], dependecies: [Dependency]) {
-        let spec = try self.loadSpec(path: self.rootSpecPath.absoluteString)
+        let spec = try wrap(
+            self.loadSpec(path: self.rootSpecPath.absoluteString),
+            message: "While loading spec at \(self.rootSpecPath.absoluteString)")
 
         var root = Dependency(pathToCurrentFile: self.rootSpecPath.absoluteString, dependecies: [:])
 
@@ -108,6 +113,7 @@ extension ReferenceExtractor {
             case let sp as [String: Any]:
                 try self.collectAllRefs(in: sp, file: file, currentDependency: &currentDependency)
             case let str as String:
+
                 guard key == "$ref" else {
                     continue
                 }
@@ -127,14 +133,15 @@ extension ReferenceExtractor {
 
 
     func readOther(filePath: String, fromFile: String, currentDependency: inout Dependency, refString: String) throws {
-        var rootUrl = self.rootSpecPath
+        guard var rootUrl = URL(string: currentDependency.pathToCurrentFile) else {
+            throw CustomError(message: "Couldn't convert \(currentDependency.pathToCurrentFile) to URI")
+        }
 
         rootUrl.deleteLastPathComponent()
 
         let resultedUrlToFileToParse = try rootUrl.absoluteString.appending(filePath).normalized()
 
         currentDependency.dependecies[refString] = resultedUrlToFileToParse
-
 
         // if we already read this file we won't read it again
         if self.readStack.contains(resultedUrlToFileToParse) {
@@ -153,7 +160,10 @@ extension ReferenceExtractor {
         // and we don't need to read in twice
         self.readStack.append(resultedUrlToFileToParse)
 
-        let res = try self.loadSpec(path: resultedUrlToFileToParse)
+        let res = try wrap(
+            self.loadSpec(path: resultedUrlToFileToParse),
+            message: "While loading spec at \(resultedUrlToFileToParse)"
+        )
 
         var newDependency = Dependency(pathToCurrentFile: resultedUrlToFileToParse, dependecies: [:])
